Azure/az-funcctions/requirements.txt
azure-functions
azure-mgmt-resource==5.1.0
azure-mgmt-security==0.3.0
azure-cosmosdb-table==1.0.6
azure-mgmt-policyinsights==0.3.1
